Ellipse points

Ellipse points

Usage

ellipse_points(par_ellipse, columns = c(1, 2), n_points = 100)

Arguments

par_ellipse

An ellipse object with ellipse parameters

columns

The number of columns of the ellipse to get the points

n_points

The number of points for the ellipses to plot

Value

A data frame with the coordinates to plot

Examples

ellipse_points(ellipse_par_example)
#>               x         y variable
#> 1   -52.1089888 2230.6113      1_2
#> 2   -51.0330167 2349.2488      1_2
#> 3   -49.6564537 2467.4061      1_2
#> 4   -47.9848427 2584.6077      1_2
#> 5   -46.0249147 2700.3814      1_2
#> 6   -43.7845617 2814.2612      1_2
#> 7   -41.2728046 2925.7884      1_2
#> 8   -38.4997576 3034.5141      1_2
#> 9   -35.4765867 3140.0003      1_2
#> 10  -32.2154650 3241.8224      1_2
#> 11  -28.7295241 3339.5703      1_2
#> 12  -25.0328006 3432.8504      1_2
#> 13  -21.1401798 3521.2872      1_2
#> 14  -17.0673360 3604.5245      1_2
#> 15  -12.8306690 3682.2272      1_2
#> 16   -8.4472385 3754.0823      1_2
#> 17   -3.9346950 3819.8006      1_2
#> 18    0.6887912 3879.1174      1_2
#> 19    5.4046028 3931.7939      1_2
#> 20   10.1937509 3977.6180      1_2
#> 21   15.0369514 4016.4051      1_2
#> 22   19.9147024 4047.9991      1_2
#> 23   24.8073629 4072.2727      1_2
#> 24   29.6952319 4089.1282      1_2
#> 25   34.5586277 4098.4978      1_2
#> 26   39.3779670 4100.3436      1_2
#> 27   44.1338441 4094.6584      1_2
#> 28   48.8071089 4081.4649      1_2
#> 29   53.3789436 4060.8163      1_2
#> 30   57.8309391 4032.7957      1_2
#> 31   62.1451688 3997.5160      1_2
#> 32   66.3042609 3955.1192      1_2
#> 33   70.2914680 3905.7761      1_2
#> 34   74.0907352 3849.6853      1_2
#> 35   77.6867642 3787.0727      1_2
#> 36   81.0650750 3718.1904      1_2
#> 37   84.2120643 3643.3157      1_2
#> 38   87.1150603 3562.7502      1_2
#> 39   89.7623737 3476.8183      1_2
#> 40   92.1433447 3385.8660      1_2
#> 41   94.2483860 3290.2595      1_2
#> 42   96.0690212 3190.3838      1_2
#> 43   97.5979194 3086.6411      1_2
#> 44   98.8289242 2979.4491      1_2
#> 45   99.7570788 2869.2393      1_2
#> 46  100.3786458 2756.4557      1_2
#> 47  100.6911224 2641.5523      1_2
#> 48  100.6932504 2524.9919      1_2
#> 49  100.3850212 2407.2437      1_2
#> 50   99.7676760 2288.7818      1_2
#> 51   98.8437005 2170.0834      1_2
#> 52   97.6168153 2051.6263      1_2
#> 53   96.0919606 1933.8876      1_2
#> 54   94.2752764 1817.3413      1_2
#> 55   92.1740780 1702.4567      1_2
#> 56   89.7968261 1589.6964      1_2
#> 57   87.1530931 1479.5145      1_2
#> 58   84.2535243 1372.3546      1_2
#> 59   81.1097953 1268.6483      1_2
#> 60   77.7345647 1168.8130      1_2
#> 61   74.1414235 1073.2509      1_2
#> 62   70.3448399  982.3466      1_2
#> 63   66.3601015  896.4664      1_2
#> 64   62.2032533  815.9558      1_2
#> 65   57.8910335  741.1393      1_2
#> 66   53.4408060  672.3179      1_2
#> 67   48.8704902  609.7688      1_2
#> 68   44.1984892  553.7440      1_2
#> 69   39.4436155  504.4689      1_2
#> 70   34.6250152  462.1420      1_2
#> 71   29.7620911  426.9337      1_2
#> 72   24.8744247  398.9859      1_2
#> 73   19.9816966  378.4109      1_2
#> 74   15.1036083  365.2918      1_2
#> 75   10.2598021  359.6813      1_2
#> 76    5.4697823  361.6019      1_2
#> 77    0.7528366  371.0461      1_2
#> 78   -3.8720415  387.9756      1_2
#> 79   -8.3862293  412.3225      1_2
#> 80  -12.7715498  443.9886      1_2
#> 81  -17.0103447  482.8464      1_2
#> 82  -21.0855460  528.7394      1_2
#> 83  -24.9807443  581.4829      1_2
#> 84  -28.6802549  640.8645      1_2
#> 85  -32.1691813  706.6451      1_2
#> 86  -35.4334748  778.5598      1_2
#> 87  -38.4599912  856.3190      1_2
#> 88  -41.2365438  939.6096      1_2
#> 89  -43.7519524 1028.0963      1_2
#> 90  -45.9960883 1121.4227      1_2
#> 91  -47.9599153 1219.2130      1_2
#> 92  -49.6355256 1321.0735      1_2
#> 93  -51.0161722 1426.5940      1_2
#> 94  -52.0962957 1535.3497      1_2
#> 95  -52.8715468 1646.9025      1_2
#> 96  -53.3388038 1760.8033      1_2
#> 97  -53.4961854 1876.5936      1_2
#> 98  -53.3430577 1993.8069      1_2
#> 99  -52.8800374 2111.9714      1_2
#> 100 -52.1089888 2230.6113      1_2
ellipse_points(ellipse_par_example, columns = c(1, 3)  )
#>              x           y variable
#> 1   -53.464245  212.242729      1_3
#> 2   -53.168285  260.959448      1_3
#> 3   -52.563137  309.479104      1_3
#> 4   -51.651236  357.606326      1_3
#> 5   -50.436254  405.147324      1_3
#> 6   -48.923085  451.910667      1_3
#> 7   -47.117821  497.708054      1_3
#> 8   -45.027731  542.355077      1_3
#> 9   -42.661231  585.671956      1_3
#> 10  -40.027851  627.484271      1_3
#> 11  -37.138193  667.623659      1_3
#> 12  -34.003895  705.928491      1_3
#> 13  -30.637575  742.244527      1_3
#> 14  -27.052790  776.425537      1_3
#> 15  -23.263974  808.333885      1_3
#> 16  -19.286383  837.841087      1_3
#> 17  -15.136033  864.828329      1_3
#> 18  -10.829637  889.186942      1_3
#> 19   -6.384535  910.818843      1_3
#> 20   -1.818626  929.636927      1_3
#> 21    2.849705  945.565422      1_3
#> 22    7.601660  958.540187      1_3
#> 23   12.418106  968.508980      1_3
#> 24   17.279647  975.431658      1_3
#> 25   22.166707  979.280347      1_3
#> 26   27.059610  980.039550      1_3
#> 27   31.938651  977.706208      1_3
#> 28   36.784187  972.289719      1_3
#> 29   41.576704  963.811892      1_3
#> 30   46.296905  952.306864      1_3
#> 31   50.925784  937.820962      1_3
#> 32   55.444702  920.412516      1_3
#> 33   59.835463  900.151624      1_3
#> 34   64.080386  877.119868      1_3
#> 35   68.162379  851.409990      1_3
#> 36   72.065006  823.125514      1_3
#> 37   75.772551  792.380331      1_3
#> 38   79.270085  759.298243      1_3
#> 39   82.543526  724.012458      1_3
#> 40   85.579693  686.665060      1_3
#> 41   88.366359  647.406434      1_3
#> 42   90.892304  606.394661      1_3
#> 43   93.147356  563.794880      1_3
#> 44   95.122437  519.778626      1_3
#> 45   96.809591  474.523136      1_3
#> 46   98.202027  428.210639      1_3
#> 47   99.294137  381.027618      1_3
#> 48  100.081523  333.164064      1_3
#> 49  100.561015  284.812705      1_3
#> 50  100.730682  236.168236      1_3
#> 51  100.589842  187.426530      1_3
#> 52  100.139061  138.783853      1_3
#> 53   99.380154   90.436073      1_3
#> 54   98.316177   42.577868      1_3
#> 55   96.951415   -4.598054      1_3
#> 56   95.291362  -50.901731      1_3
#> 57   93.342704  -96.146714      1_3
#> 58   91.113287 -140.150820      1_3
#> 59   88.612088 -182.736857      1_3
#> 60   85.849178 -223.733348      1_3
#> 61   82.835682 -262.975215      1_3
#> 62   79.583736 -300.304443      1_3
#> 63   76.106433 -335.570721      1_3
#> 64   72.417775 -368.632045      1_3
#> 65   68.532616 -399.355288      1_3
#> 66   64.466598 -427.616738      1_3
#> 67   60.236096 -453.302597      1_3
#> 68   55.858143 -476.309437      1_3
#> 69   51.350368 -496.544617      1_3
#> 70   46.730922 -513.926658      1_3
#> 71   42.018406 -528.385567      1_3
#> 72   37.231796 -539.863125      1_3
#> 73   32.390365 -548.313114      1_3
#> 74   27.513609 -553.701511      1_3
#> 75   22.621164 -556.006617      1_3
#> 76   17.732731 -555.219151      1_3
#> 77   12.867994 -551.342284      1_3
#> 78    8.046540 -544.391627      1_3
#> 79    3.287785 -534.395166      1_3
#> 80   -1.389110 -521.393156      1_3
#> 81   -5.965312 -505.437949      1_3
#> 82  -10.422396 -486.593793      1_3
#> 83  -14.742412 -464.936565      1_3
#> 84  -18.907968 -440.553472      1_3
#> 85  -22.902289 -413.542696      1_3
#> 86  -26.709291 -384.013000      1_3
#> 87  -30.313646 -352.083289      1_3
#> 88  -33.700839 -317.882134      1_3
#> 89  -36.857231 -281.547249      1_3
#> 90  -39.770114 -243.224944      1_3
#> 91  -42.427758 -203.069528      1_3
#> 92  -44.819461 -161.242692      1_3
#> 93  -46.935592 -117.912860      1_3
#> 94  -48.767632  -73.254505      1_3
#> 95  -50.308203  -27.447449      1_3
#> 96  -51.551101   19.323857      1_3
#> 97  -52.491323   66.871082      1_3
#> 98  -53.125081  115.002771      1_3
#> 99  -53.449824  163.525115      1_3
#> 100 -53.464245  212.242729      1_3