get_env_var Function to extract environmental values from presence points

get_env_var Function to extract environmental values from presence points

Usage

get_env_var(df, env)

Arguments

df

A data frame with presence points

env

A raster with environmental layers. Must have a projection string

Examples

stack_path <- system.file("extdata", "stack_1_12_crop.rds", package="nicher")
stack_1_12 <- terra::unwrap(readr::read_rds(stack_path))
get_env_var(rawSpOccPnts, stack_1_12)
#>      bio1WH bio12WH
#> 1  20.73091    2075
#> 2  22.41057    1630
#> 3  22.25763    2203
#> 4  19.31113    1897
#> 5  23.88845    3060
#> 6  23.12673    2032
#> 7  22.76155    1337
#> 8  20.18799    2251
#> 9  19.77320    1529
#> 10 21.94147    2183
#> 11 16.30925    1144
#> 12 22.55396    1250
#> 13 19.15915    1923
#> 14 21.92423    2415
#> 15 18.30095    2529
#> 16 20.68048    1930
#> 17 24.19290    1525
#> 18 23.83894    1282
#> 19 16.43809    1092
#> 20 22.74709    1398
#> 21 22.03944    1029
#> 22 27.67044    2215
#> 23 24.83927    2112
#> 24 25.83965    3651
#> 25 14.50923    1563
#> 26 17.58682     747
#> 27 23.31386     862
#> 28 26.99096    1306
#> 29 26.65801    1157
#> 30 17.36748    2033
#> 31 23.38448    3589
#> 32 16.26900     807
#> 33 21.52402    1828
#> 34 20.87229    2811
#> 35 21.63052    1224
#> 36 19.35501    1699
#> 37 22.56491    1887
#> 38 20.60624    1391
#> 39 23.61762    1006
#> 40 25.48557    1417
#> 41 25.11342    3426
#> 42 21.38318    1287
#> 43 18.72291    1286
#> 44 16.02533    2420
#> 45 26.18298    2968
#> 46 20.35045    2072
#> 47 20.91000    3461
#> 48 19.00921    2590
#> 49 19.21876    2919
#> 50 19.01612    1788
#> 51 22.30757    1826
#> 52 23.57381    1553
#> 53 22.95685    1722
#> 54 21.33614    1322
#> 55 20.42602    1561
#> 56 23.34570    1265
#> 57 20.15109    1454
#> 58 19.35501    1699
#> 59 20.91557    1856
#> 60 22.42379    1726
#> 61 21.98492    1590
#> 62 21.52402    1828
#> 63 23.17664    1832
#> 64 24.98680    2832
#> 65 20.68048    1930
#> 66 16.06038    1210
#> 67 24.79048    2453
#> 68 23.95322    1985
#> 69 17.48504    2404
#> 70 17.57528    1206
#> 71 22.47643    2953
#> 72 20.31801    1826
#> 73 20.86426    2161