get_suitability Function that calculates the log(suitability)

get_suitability Function that calculates the log(suitability)

Usage

get_suitability(df, el_pars)

Arguments

df

A data frame with environmental information

el_pars

A list with ellipse parameter. Contains a vector of means and a covariance matrix. Could came from a MLE after optimization or calculated from only presence data.

Value

A vector with suitability

Examples

pars <- get_negloglike_optimr_par(head(spOccPnts, 10), samMPts, fastmode = TRUE, itnmax = 1)
#> fastmode
#> TRUE
#> Optimization starting 
#> Methods used:
#> L-BFGS-B
#> Optimization finished. 
#> L-BFGS-B is the returned method
get_suitability(spOccPnts, pars)
#>  [1] 0.915875581 0.154225911 0.427715315 0.290628590 0.050188570 0.159792513
#>  [7] 0.061976806 0.416796424 0.291835121 0.576855772 0.025699160 0.058407033
#> [13] 0.244341232 0.407305603 0.048101447 0.923920848 0.034400678 0.029760140
#> [19] 0.026713939 0.071074101 0.046978551 0.008998038 0.039162020 0.013644548
#> [25] 0.010814196 0.031155681 0.019892173 0.006991403 0.007057324 0.045642468
#> [31] 0.024224096 0.020559266 0.564828819 0.116010590 0.090204579 0.290292247
#> [37] 0.231116703 0.216815795 0.021772211 0.014853853 0.021038021 0.120428196
#> [43] 0.107397875 0.014498604 0.019099832 0.714373855 0.025912480 0.069762645
#> [49] 0.040941164 0.220167995 0.266491588 0.053930837 0.117262231 0.136131953
#> [55] 0.377516643 0.038176302 0.267232896 0.290292247 0.838739049 0.190750699
#> [61] 0.197185917 0.564828819 0.116714165 0.037247282 0.923920848 0.023196721
#> [67] 0.047217310 0.072230595 0.033602308 0.052330830 0.092084498 0.703525319
#> [73] 0.841132058